fix(devices): improve printer spooler, WIA scanner resolution, device discovery and bump to v1.2.3

+++ Services/BlankPageDetector.cs
@@ -114,8 +114,8 @@ namespace DocConverter.Services
                     {
                         var readSettings = new MagickReadSettings
                         {
-                            FrameIndex = i,
-                            FrameCount = 1,
+                            FrameIndex = (uint)i,
+                            FrameCount = 1U,
                             Density = new Density(72)
                         };
